refactor: #138 불필요한 로직 제거
리펙토링 후 불필요해진 로직을 제거하였다.
