+++ src/main/java/com/aliens/friendship/domain/auth/dto/TokenDto.java
@@ -1,4 +1,4 @@
-package com.aliens.friendship.domain.jwt.domain.dto;
+package com.aliens.friendship.domain.auth.dto;
 
 import lombok.*;
