+++ src/main/java/com/aliens/friendship/domain/auth/model/principal/UserPrincipal.java
@@ -1,9 +1,8 @@
-package com.aliens.friendship.domain.auth.principal;
+package com.aliens.friendship.domain.auth.model.principal;
 
 import com.aliens.friendship.domain.member.domain.Member;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import java.util.Collection;
