--- src/main/java/com/aliens/friendship/domain/jwt/domain/LogoutAccessToken.java
@@ -1,32 +1,0 @@
-package com.aliens.friendship.domain.jwt.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.springframework.data.redis.core.RedisHash;
-import org.springframework.data.redis.core.TimeToLive;
-
-import javax.persistence.Id;
-
-@Getter
-@RedisHash("logoutAccessToken")
-@AllArgsConstructor
-@Builder
-public class LogoutAccessToken {
-
-    @Id
-    private String id;
-
-    private String username;
-
-    @TimeToLive
-    private Long expiration;
-
-    public static LogoutAccessToken of(String accessToken, String username, Long remainingMilliSeconds) {
-        return LogoutAccessToken.builder()
-                .id(accessToken)
-                .username(username)
-                .expiration(remainingMilliSeconds / 1000)
-                .build();
-    }
-}
