--- src/main/java/com/aliens/friendship/domain/jwt/domain/dto/LoginDto.java
@@ -1,12 +1,0 @@
-package com.aliens.friendship.domain.jwt.domain.dto;
-
-
-import lombok.*;
-
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor @Builder
-public class LoginDto {
-
-    private String email;
-    private String password;
-}
