--- src/main/java/com/aliens/friendship/domain/jwt/domain/dto/MemberInfo.java
@@ -1,10 +1,0 @@
-package com.aliens.friendship.domain.jwt.domain.dto;
-import lombok.*;
-
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor @Builder
-public class MemberInfo {
-
-    private String email;
-    private String username;
-}
