--- src/main/java/com/aliens/friendship/domain/jwt/exception/JWTExceptionHandler.java
@@ -1,44 +1,0 @@
-package com.aliens.friendship.domain.jwt.exception;
-
-import com.aliens.friendship.global.error.ErrorResponse;
-import com.aliens.friendship.global.error.ExceptionCode;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class JWTExceptionHandler {
-
-    /**
-     * TokenException handling (Custom Exception)
-     */
-    @ExceptionHandler(TokenException.class)
-    protected ResponseEntity<ErrorResponse> handleTokenException(
-            TokenException e
-    ) {
-        ExceptionCode exceptionCode = e.getExceptionCode();
-        log.error("{}", e.getMessage());
-        return new ResponseEntity<>(
-                ErrorResponse.of(exceptionCode, exceptionCode.getMessage()),
-                HttpStatus.valueOf(exceptionCode.getHttpStatus().value())
-        );
-    }
-
-    /**
-     * RefreshTokenNotFoundException handling (Custom Exception)
-     */
-    @ExceptionHandler(RefreshTokenNotFoundException.class)
-    protected ResponseEntity<ErrorResponse> handleRefreshTokenNotFoundException(
-            RefreshTokenNotFoundException e
-    ) {
-        ExceptionCode exceptionCode = e.getExceptionCode();
-        log.error("{}", e.getMessage());
-        return new ResponseEntity<>(
-                ErrorResponse.of(exceptionCode, exceptionCode.getMessage()),
-                HttpStatus.valueOf(exceptionCode.getHttpStatus().value())
-        );
-    }
-}
