--- src/main/java/com/aliens/friendship/domain/jwt/repository/LogoutAccessTokenRedisRepository.java
@@ -1,7 +1,0 @@
-package com.aliens.friendship.domain.jwt.repository;
-
-import com.aliens.friendship.domain.jwt.domain.LogoutAccessToken;
-import org.springframework.data.repository.CrudRepository;
-
-public interface LogoutAccessTokenRedisRepository extends CrudRepository<LogoutAccessToken, String> {
-}
