--- src/main/java/com/aliens/friendship/domain/jwt/util/JwtTokenUtil.java
@@ -1,78 +1,0 @@
-package com.aliens.friendship.domain.jwt.util;
-
-import com.aliens.friendship.global.config.jwt.JwtExpirationEnums;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.security.Keys;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import java.nio.charset.StandardCharsets;
-import java.security.Key;
-import java.util.Date;
-
-
-@Slf4j
-@Component
-public class JwtTokenUtil {
-
-    @Value("${spring.jwt.secret}")
-    private String SECRET_KEY;
-
-    public Claims extractAllClaims(String token) {
-        return Jwts.parserBuilder()
-                .setSigningKey(getSigningKey(SECRET_KEY))
-                .build()
-                .parseClaimsJws(token)
-                .getBody();
-    }
-
-    public String getEmail(String token) {
-        return extractAllClaims(token).get("email", String.class);
-    }
-
-    private Key getSigningKey(String secretKey) {
-        byte[] keyBytes = secretKey.getBytes(StandardCharsets.UTF_8);
-        return Keys.hmacShaKeyFor(keyBytes);
-    }
-
-    public Boolean isTokenExpired(String token) {
-        Date expiration = extractAllClaims(token).getExpiration();
-        return expiration.before(new Date());
-    }
-
-    public String generateAccessToken(String email) {
-        return doGenerateToken(email, JwtExpirationEnums.ACCESS_TOKEN_EXPIRATION_TIME.getValue());
-    }
-
-    public String generateRefreshToken(String email) {
-        return doGenerateToken(email, JwtExpirationEnums.REFRESH_TOKEN_EXPIRATION_TIME.getValue());
-    }
-
-    private String doGenerateToken(String email, long expireTime) {
-        Claims claims = Jwts.claims();
-        claims.put("email", email);
-
-        return Jwts.builder()
-                .setClaims(claims)
-                .setIssuedAt(new Date(System.currentTimeMillis()))
-                .setExpiration(new Date(System.currentTimeMillis() + expireTime))
-                .signWith(getSigningKey(SECRET_KEY), SignatureAlgorithm.HS256)
-                .compact();
-    }
-
-    public Boolean validateToken(String token, UserDetails userDetails) {
-        String username = getEmail(token);
-        return username.equals(userDetails.getUsername())
-                && !isTokenExpired(token);
-    }
-
-    public long getRemainMilliSeconds(String token) {
-        Date expiration = extractAllClaims(token).getExpiration();
-        Date now = new Date();
-        return expiration.getTime() - now.getTime();
-    }
-}
