+++ src/main/java/com/aliens/friendship/domain/member/controller/APIController.java
@@ -1,13 +1,14 @@
 package com.aliens.friendship.domain.member.controller;
 
-import com.aliens.friendship.domain.jwt.domain.dto.TokenDto;
-import com.aliens.friendship.domain.jwt.util.JwtTokenUtil;
 import com.aliens.friendship.domain.member.controller.dto.JoinDto;
 import com.aliens.friendship.domain.member.service.MemberService;
 import com.aliens.friendship.global.response.CommonResult;
 import com.aliens.friendship.global.response.ResponseService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import static org.springframework.http.HttpStatus.OK;
 
@@ -16,7 +17,6 @@ import static org.springframework.http.HttpStatus.OK;
 public class APIController {
 
     private final MemberService memberService;
-    private final JwtTokenUtil jwtTokenUtil;
     private final ResponseService responseService;
 
     @GetMapping("/health")
@@ -35,19 +35,6 @@ public class APIController {
         return responseService.getSuccessResult(
                 OK.value(),
                 "회원가입 완료"
-        );
-    }
-
-    @PostMapping("/logout")
-    public CommonResult logout(
-            @RequestHeader("Authorization") String accessToken,
-            @RequestHeader("RefreshToken") String refreshToken
-    ) {
-        String email = jwtTokenUtil.getEmail(memberService.resolveToken(accessToken));
-        memberService.logout(TokenDto.of(accessToken, refreshToken), email);
-        return responseService.getSuccessResult(
-                OK.value(),
-                "성공적으로 로그아웃되었습니다."
         );
     }
 }
